Implement a wasm3 image
